Remove auto-casting to NetworkEndpoint#getContext() method's result type.

+++ hekate-core/src/test/java/io/hekate/javadoc/network/NetworkServiceJavadocTest.java
@@ -53,7 +53,7 @@ public class NetworkServiceJavadocTest extends HekateNodeTestBase {
 
             System.out.println("Message from client: " + msg);
 
-            AtomicInteger counter = from.getContext();
+            AtomicInteger counter = (AtomicInteger)from.getContext();
 
             // Send reply.
             from.send(msg + " processed (total=" + counter.incrementAndGet() + ')');
